+++ packages/Fakturoid/config/config.yaml
@@ -1,0 +1,12 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+        autoconfigure: true
+
+    Pehapkari\Fakturoid\:
+        resource: '../src'
+        exclude: '../src/ValueObject/*'
+
+    Pehapkari\Fakturoid\Guzzle\FakturoidClient:
+        factory: ['@Pehapkari\Fakturoid\Guzzle\FakturoidClientFactory', 'create']
